add testone option to makefile
Add an option to test only one configuration and raise an error
in case of a problem.

Use one_o as a makefile variable because o is used in build to
handle the extension of the .o files.

Use one_cp0 as a makefile variable because cp0 is used in build to
as a numeric value instead of a boolean value.