Fixes and small improvements for type recovery.
Removed counter field from prelex, using the operand field instead to
provide the index into the fxmap.  This follows other uses within the
compiler
where we use the operand field as a little place for state that is used
within a single pass.  This has a few advantages.  First, it keeps the
record a
little smaller.  Second, it means that the prelex numbering can start
from 0
for each compilation unit, which should help keep the numbers for the
fxmap a
bit smaller in longer running sessions with multiple calls to the
compiler.
Finally, it avoids adding to the burden of the tc-mutex, since within
the pass
it is safe for us to set the prelexes, since only the instance of the
pass
holding this block of code has a handle on it.  As part of this change
prelex-counter is now defined in cptypes and the operand is cleared
after the
variables go out of scope.
  base-lang.ss

Fixed the highest-set-bit function in fxmap so that it will work in the
32-bit
versions of Chez Scheme.  The fxsrl by 32 raises an exception, and was
leading
to tests to fail in 32-bit mode.
  fxmap.ss

Restructured predicate-implies? so that it uses committed choice instead
of
uncommitted choice in comparing x and y.  Basically, this means, instead
of
doing:
(or
  (and (predicate-1? x) (predicate-1? y) ---)
  (and (predicate-2? x) (predicate-2? y) ---)
  ...)
we now do:
(cond
  [(predicate-1? x) (and (predicate-1? y) ---)]
  [(predicate-2? x) (and (predicate-2? y) ---)]
  ...)
This avoids running predicates on x that we know will fail because an
earlier
predicate matches, generally getting out of the predicate faster.  This
did
require a little restructuring, because in some cases x was dominant and
in
other cases y was dominant.  This is now restructured with y dominate,
after
the eq? and x 'bottom check.

Replaced let-values calls with cata-morphism syntax, including removal
of maps
that built up a list of values that then needed to be separated out with
(map car ...) (map cadr ...) etc. calls.  This avoid building up
structures we
don't need, since the nanopass framework will generate a mutltivalued
let for
these situations.

The if clause in cptypes/raw now uses types1 (the result of the
recursive call
on e1) in place of the incoming types clause when processing the e2 or
e3
expressions in the cases where e1 is known statically to produce either
a false
or non-false value.

Fixed a bug with directly-applied variable arity lambda.  The original
code
marked all directly-applied variable arity lambda's as producing bottom,
because it was chacking for the interface to be equal to the number of
arguments.  However, variable arity functions are represented with a
negative
number.  For instance, the original code would transform the expression:
(begin
  ((lambda (a . b) (set! t (cons* b a t))) 'a 'b 'c)
  t)
to
((lambda (a . b) (set! t (cons* b a t))) 'a 'b 'c)
anticipating that the call would raise an error, however, it is a
perfectly
valid (if some what unusual) expression.  I tried to come up with a test
for
this, however, without building something fairly complicated, it is
difficult
to get past cp0 without cp0 turning it into something like:
(let ([b (list 'b 'c)])
  (set! t (cons* b 'a t))
  t)

Fixed make-time, time-second-set!, and time-second to indicate that
second can
be an exact-integer, since it is not really restricted to the fixnum
range (and
if fact we even test this fact in the mats on 32-bit machines).
  primdata.ss

Changed check of prelex-was-assigned (which is not reliably on the input
to any
give pass) with prelex-assigned, which should always have an accurate,
if
conservative, value in it.

Added enable-type-recovery parameter to allow the type recover to be
turned on
and off, and added cptype to the cp0 not run path that runs cpletrec, so
that
cptypes can be run independent of cp0.  This is helpful for testing and
allows
us to benefit from type recovery, even in cases where we do not want cp0
to
perform any inlining.
  compile.ss, front.ss, primdata.ss

Stylistic changes, mostly for consistency with other parts of the
compiler,
though I'm not married to these changes if you'd really prefer to keep
things
the way the are.
1. clauses of define-record type now use parenthesis instead of square
brackets.
2. indented by 2 spaces where things were only indented by one space
3. define, let, define-pass, nanopass pass productions clauses, now use
   parenthesis for outer markers instead of square brackets.
  fxmap.ss,